+++ CHANGELOG.rst
@@ -5,8 +5,12 @@ Changelog for pytest-find-dependencies
 Unreleased
 ----------
 
-Infrastructure:
-~~~~~~~~~~~~~~~
+Changes:
+~~~~~~~~
+* set exit status to failed if dependent tests are found (see `#3`_)
+
+Fixes:
+~~~~~~
 * ignore warnings in tests where they may change the outcome (see `#2`_)
 
 `0.4.1`_ (2021-12-26)
@@ -79,3 +83,4 @@ New Features
 .. _`0.4.1`: https://pypi.org/project/pytest-find-dependencies/0.4.1/
 .. _`#1`: https://github.com/mrbean-bremen/pytest-find-dependencies/issues/1
 .. _`#2`: https://github.com/mrbean-bremen/pytest-find-dependencies/issues/2
+.. _`#3`: https://github.com/mrbean-bremen/pytest-find-dependencies/issues/3